Show formatted & encoded quote on solve view
Solve view correctly gets & displays the formatted, encoded quote and
author
