+++ src/components/App/App.js
@@ -34,8 +34,6 @@ class App extends Component {
   render () {
     const { msgAlerts, user } = this.state
 
-    console.log('user', user)
-
     return (
       <Fragment>
         <Header user={user} />
